refactor(core): split up components
