+++ packages/sanity/src/core/form/studio/tree-editing/components/index.ts
@@ -1,5 +1,5 @@
 export * from './breadcrumbs'
+export * from './layout'
 export * from './search'
+export * from './tree-menu'
 export * from './TreeEditingDialog'
-export * from './TreeEditingLayout'
-export * from './TreeEditingMenu'
