+++ packages/sanity/src/core/form/studio/tree-editing/components/layout/index.ts
@@ -1,0 +1,1 @@
+export * from './TreeEditingLayout'
