+++ packages/sanity/src/core/form/studio/tree-editing/components/layout/resizer/Resizer.tsx
