+++ packages/sanity/src/core/form/studio/tree-editing/components/tree-menu/index.ts
@@ -1,0 +1,1 @@
+export * from './TreeEditingMenu'
