+++ packages/sanity/src/core/form/studio/tree-editing/hooks/index.ts
@@ -1,1 +1,2 @@
+export * from './useSearchableList'
 export * from './useValuePreviewWithFallback'
